requirements.txt
Django==2.2.1
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
future==0.17.1
PyJWT==1.7.1
pytz==2019.1
serializers==0.2.4
sqlparse==0.3.0
